fix(sentinel): ensure sentinel re-observes after items render
The intersection observer will re-attach the target, to fix a case occurring when an itemLength update
is too small and has the sentinel re-render within the intersecting area.
This prevented the intersection event from happening either until the intersection changed again or alltogether
when using `awaitMore` due to the `awaitIntersection` set fag.